trends-analysis/requirements.txt
blinker==1.7.0; python_version >= '3.8'
boto3==1.33.11; python_version >= '3.7'
botocore==1.33.11; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
flask==3.0.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
importlib-metadata==7.0.0; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
joblib==1.3.2; python_version >= '3.7'
lxml==4.9.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==2.1.3; python_version >= '3.7'
numpy==1.26.2; python_version >= '3.9'
pandas==2.1.4; python_version >= '3.9'
psycopg2==2.9.9; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytrends==4.9.2; python_version >= '3.7'
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
s3transfer==0.8.2; python_version >= '3.7'
scikit-learn==1.3.2; python_version >= '3.8'
scipy==1.11.4; python_version >= '3.9'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.23; python_version >= '3.7'
threadpoolctl==3.2.0; python_version >= '3.8'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2023.3; python_version >= '2'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
waitress==2.1.2; python_full_version >= '3.7.0'
werkzeug==3.0.1; python_version >= '3.8'
zipp==3.17.0; python_version >= '3.8'
